pyproject.toml
[tool.poetry]
name = "flask-zs-template"
version = "0.1.0"
authors = ["codeif <me@codeif.com>"]
description = "A flask-zs example"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[[tool.poetry.source]]
name = "pypi.org"
url = "https://pypi.org/simple"

[tool.poetry.dependencies]
python = "^3.7"
pymysql = "^0.9.3"
flask-sqlalchemy = "^2.4.1"
flask-celery = "^2.4.3"
flask-redis = "^0.4.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

# [build-system]
# requires = ["poetry>=0.12"]
# build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''


# Build system information below.
# NOTE: You don't need this in your own Black configuration.

[build-system]
requires = ["setuptools>=41.0", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"
